+++ test/test_screen_test.dart
@@ -46,6 +46,9 @@ class MockTestUploader implements BaseUploader {
   String? get proxyUrl => null;
 
   @override
+  String? get instanceDescription => null;
+
+  @override
   ProviderMetadata get metadata => ProviderMetadata(
         allowedMimeTypes: isImageProvider ? {'image/*'} : null,
       );