+++ test/upload_provider_test.dart
@@ -69,6 +69,9 @@ class MockBaseUploader implements BaseUploader {
   String? get proxyUrl => null;
 
   @override
+  String? get instanceDescription => null;
+
+  @override
   ProviderMetadata get metadata => _metadata;
 
   @override