+++ test/upload_screen_test.dart
@@ -75,6 +75,9 @@ class MockUploader implements BaseUploader {
   String? get proxyUrl => null;
 
   @override
+  String? get instanceDescription => null;
+
+  @override
   ProviderMetadata get metadata => ProviderMetadata(
         allowedMimeTypes: isImageProvider ? {'image/*'} : null,
       );